Made daily forecast functional
In the displayForecast function:

Replaced forecast.forEach with forecast.slice(1, 7).forEach to start the forecast from the next day (index 1) and limit it to the next 6 days.
Removed the index parameter from the forEach function since it is not used.
Updated the forecast day calculation by calling formatDay(forecastDay.dt).
In the displayWeatherData function:

Updated the getDailyForecast function call to pass response.data.coord as the argument.
In the variable declaration section:

Updated the apiKey value to the correct value for the OpenWeather API.
Removed the redundant code at the end that called the displayForecast function.